selectlist: add <listbox> element
In the new selectlist architecture, the <listbox> element is the
replacement for the behavior=listbox, slot=listbox, and popover=auto
attributes. When a <listbox> is specified as a child of a <selectlist>,
then it will be slotted into the listbox slot so that the <listbox> and
all of its children will be rendered as the popup of options.

Bug: 1121840
Change-Id: I16b1629f418b7ee60027d7b9d727a370d3ffa3dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4807178
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192026}
NOKEYCHECK=True
GitOrigin-RevId: 43a6b6eeba8dcb01afa9a3a94b74371596287e95