Allow CSSParserTokenStream::Restore during Boundary
There is no technical reason to disallow this, it was just omitted
from the first CL to keep it simple. This CL lifts the restriction,
and at the same time adds DCHECKs to ensure that the boundary is
the same during Save and Restore.

This makes Save/Restore useful in way more situations, and should
allow progress on Issue 1361240.

Bug: 1361240, 1427259
Change-Id: I1e74cd59c8e42f1ac3cda13bbceb332329fbd1d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4833759
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192073}
NOKEYCHECK=True
GitOrigin-RevId: 22e4cce51b19b4f08d3c3f5b8d15add4608c0254