Plumbling cache_storage access to SW subresource loader.
This is a part of making the cache storage API accessible from the
ServiceWorkerSubresourceLoader upon the static routing API rule
(https://github.com/yoshisatoyanagisawa/service-worker-static-routing-api).

This CL makes a connection to the cache storage, and makes it available
in the subresource loader.

Bug: 1371756
Change-Id: If3c617872896e96dbaa7771eb56cdaabfc163d8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4825803
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192268}
NOKEYCHECK=True
GitOrigin-RevId: 5073c658f8c9fe92434292a982327855d5280c9a