[WAAPI] Do not cancel an animation that has its effect reset.
When an effect is associated with an animation, it resets the effect
on the animation it was previously associated with. We were handling
the handoff correctly in Animation::setEffect, but not in the
constructor.  The previous animation should not get canceled. After
resetting the effect, the effect end time becomes zero and the old
animation should be in the finished rather than idle play state. This
change has no visual effect as even if the old animation had a fill
mode set to 'forward' or "both', this fill mode will also get reset
when the effect is reset. A finished animation without a fill mode
is inactive. The distinction is only really observable
via web-animation API calls. Likely why it took this many years to
notice the issue.

Manually verified the new test on Firefox nightly as well.  Reportedly,
webkit has the correct behavior as well, though not verified.

Bug: 1474974
Change-Id: I1c61ff32e3ac9492a022dec0af22f76525fe01fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4833271
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192410}
NOKEYCHECK=True
GitOrigin-RevId: b2b2fb09d01b7fe0e1d081c92706a81bba3ae7e7