Always readjust the gradient radii for (CSS) radial gradients
If the stop offset range is zero-length/degenerate, it will be clipped
to the range [0, 1], but the radii will not be readjusted to match.
Since for the non-repeating case a stop with an offset at > 1 (100%)
will actually affect the rendered output the radii needs to be
readjusted.

Fixed: 1478473
Change-Id: Idd816865b06978e5e0ce37c83b765b708c77b7df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4833638
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Auto-Submit: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192440}
NOKEYCHECK=True
GitOrigin-RevId: e4961c56bf5acc4cabf19d7084e89e08b7b192b3