[css-nesting-ident] Check for property validity earlier
When attempting to parse a declaration, we do things in the following
order, ignoring whitespace:

 1. Consume an ident. (Normally a property name).
 2. Consume a semicolon.
 3. Consume the value tokens. This means scanning until the first
    top-level semicolon. Actual value-parsing happens later.
 4. Check if the ident from (1) is a valid property.
 5. Parse the value.

The problem is that we're checking for property validity (4) *after*
consuming the value tokens (3). This is bad for nested contexts,
because (3) is potentially a lot of wasted tokenization work if
we're going to restart the stream due to a property being invalid (4).

Therefore, this CL refactors ConsumeDeclaration such that
ConsumeValue generally happens *after* we check for property
validity.

On my machine, with CSSNestingIdent set to status:stable,
the included performance test takes around 13ms with the optimization
and 495ms without the optimization. The important part here is that
there's a clear signal in the case of a regression.

The style perftest (style_perftest.cc) seems neutral for this CL.

Bug: 1427259
Change-Id: Ifbb0c622ab8e4b01b6f8cdb4201a26109390028b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4807782
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1192938}
NOKEYCHECK=True
GitOrigin-RevId: 21c71e0bc787f0aad6de2d4c343f3428ad6c5a46