[css-nesting-ident] Handle restarts in StyleSheetHandler
Since declarations that fail to parse are now retried as nested style
rules, we can end up with CSSParserObserver events that claim that we
both have a failed declaration and a valid style rule at the same
offset. This is technically correct, but StyleSheetHandler needs a
better interpretation of this to stay in a consistent state.

In short, we now remove the CSSPropertySourceData produced by
the conflicting ObserveProperty event if a valid style rule is
seen in the same offset. This means that:

 - Token sequences that fail as declarations, but pass as
   style rules will appear as style rules.
 - Token sequences that fail as declarations, and that also fail
   as style rules will will appear as declarations (like before).

For token sequences that pass as declarations, there is of course
no difference, because we never retry parsing as style rules for
those sequences.

Bug: 1427259
Change-Id: I59377aa76284f1802c9b340767f1f25b556eec21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4835894
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193026}
NOKEYCHECK=True
GitOrigin-RevId: 9045b958a913f0da95cdd66b81c654b65e5288f2