Fix crash in unrar when trying to get password
When opening an archive with encrypted metadata, unrar expects to have a
password. If one is not provided, it will throw an exception. While
Chromium has disabled the exception throwing, we've replaced it with a
CHECK(false), which still crashes the process.

An ideal solution would be to propagate the missing password to the
calling code, but that requires significant modifications to third_party
code. Instead this CL provides a dummy password (the letter x) when
there is no password provided.

Bug: 1478286
Change-Id: Ic9f0aa6ad8b372fde76ebe1ea8f36483e878e645
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4841883
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193114}
NOKEYCHECK=True
GitOrigin-RevId: a4fde636b5d071cd0ea502f59361a26d6136db90