Get rid of canonicalize in wpt color parsing tests
This function has mostly been removed and made remaining tests _less_
clear, not more so.

For example, it was hiding the fact that the tests were mistakenly
assuming a range of [0,100] for lightness in oklch. The actual range
is [0,1] and inputs and expectations have been adapted here.

Change-Id: Id79d74d24b44519af101d76c9ba10ac71312da8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4847548
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193131}
NOKEYCHECK=True
GitOrigin-RevId: cfad9b5885696303b76bfa6a3551a2c540686f26