Enable feature flag IncreaseTileMemorySizeProportionally for all platforms
This new policy was applied to Mac only. After enabling it for a
week, it looks like it's safe to apply it to all platforms except
for Android which has its own policy. GetGpuMemoryPolicy() now
returns a ManagedMemoryPolicy according to the desktop resolution.

Bug: 1450807
Change-Id: I007abbdef9cc59cd22c5ddde9b2f7e223e74428c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4833846
Reviewed-by: Maggie Chen <magchen@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193211}
NOKEYCHECK=True
GitOrigin-RevId: b0556b75262eb84cda405c92fd8b7549373ba104