Rename NestingIdentProperty.html -> NestingIdentNonProperty.html
This new name reflects the fact that we're testing using
"not-a-property". I'd like to use the name NestingIdentProperty.html
for a future test where we *do* use a supported property name.

Also remove the &-part of the test. It has no effect on the result.

Bug: 1427259
Change-Id: Ieb6349a687801522b2c8efef0b798b9668450622
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4847452
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193525}
NOKEYCHECK=True
GitOrigin-RevId: a31aa0e6c7ac1a25c4116874ce87b57230d4522f