jni_registration_generator: Emit newline
Two different warnings in `_GenerateStubsAndAssert()` get swallowed up
when building from a TTY. This is not a problem in CI, since ninja
appears not to clobber the newline-less emission when the output is not
a TTY. However, when run "interactively," ninja's status line will mask
the output from this function, making the warning message illegible.

To fix, we simply emit a newline at the end of the message.

This CL also trivially adjusts the function so not to join more content
to `warning_message`, writing it straight to `sys.stderr`.

Fixed: 1479282
Change-Id: I953670b64d1fbf79c437f372ec4c270fc2663d6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4845956
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193553}
NOKEYCHECK=True
GitOrigin-RevId: 22e4b8c9238c6fb52c2afe67cb4f0f6411dc92b9