Split IsPreviewed from IsAutofilled in FormControlElements
IsAutofilled() used to return true if the field is filled or in a
preview state. This is confusing given the name of the function.

This CL adds a new IsPreviewed() function and splits the logic from
IsAutofilled().

Details about the bigger picture can be found on
https://docs.google.com/document/d/1DOC1l8VV1Jvo2T_Dju5nrt2a0eqkgMjwqQoY8LgRhOE/edit?resourcekey=0-O8xPiVlLmCSLO9IC0OC0Nw

Bug: 1472209
Change-Id: I4779a9a6072bed0e1cfe858a390bccb57b131892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4835908
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Dominic Battre <battre@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1193606}
NOKEYCHECK=True
GitOrigin-RevId: aeaeb0b45f51b22600c0c4fa9d0ae9ec02c5525c