Check http protocol when emitting resource timing entry
Currently when a resource load via file protocol fails, a resource
timing is emitted. We should only emit resource timing for resources
that are fetched over http(s) protocols or from web bundles.

This CL is to fix this issue. The  ResourceResponse::ShouldPopulateResourceTiming() method used for
checking protocol when load is successful checks if the
ResourceResponse::current_request_url_ is HTTP or the web bundle url
is valid.

However when the load fails, the ResourceResponse::current_request_url_
could be null, so we use ResourceRequest::url_, since we are only
checking for protocols.

Bug: 1472922
Change-Id: I674b2041f1911e9c6c04863dc38fc64a28be27af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4799574
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193611}
NOKEYCHECK=True
GitOrigin-RevId: 8646d34d410caafb9a8ce595cadfd9a0b1401a9b