[GridNG] Omit `auto` sizes for `grid-template-*` serialization
Per https://drafts.csswg.org/cssom/#serializing-css-values "If component values can be omitted or replaced with a shorter representation without changing the meaning of the value, omit/replace them."

This means that when serializing `grid-template-*`, we should always omit `auto` sizes in `grid-template-*` serialization as it is the
default value, so it will always be a shorter representation
without changing the value. This should happen regardless of whether
it was specified by the author.

This also means that the final test case expectation in
grid-template-shorthand-valid.html is incorrect, which lead me all over the place on early iterations of this CL. The above
CSSWG quote and the fact that Safari and Firefox fail for the same
reason has me confident that the assertion there is incorrect. I
will update the test to the correct behavior in a follow-up CL.

Bug: 1028283
Change-Id: Idbb8f9bac6ad0d0d633a5d5668ba689af1064eb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4824011
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193688}
NOKEYCHECK=True
GitOrigin-RevId: 7856cfaecf279722008bb6612461b6e48e934fa1