[GridNG] Update test expectation for grid-template-shorthand-valid.html
Per https://drafts.csswg.org/cssom/#serializing-css-values "If component
values can be omitted or replaced with a shorter representation without
changing the meaning of the value, omit/replace them."

This means that when serializing `grid-template-*`, we should always
omit `auto` sizes in `grid-template-*` serialization as it is the
default value, so it will always be a shorter representation
without changing the value. This should happen regardless of whether
it was specified by the author.

This means that the final test case expectation in
grid-template-shorthand-valid.html is incorrect. It is also failing in
Blink/Gecko/WebKit for this exact reason, so this CL updates the test's
expected behavior to omit `auto`. After this change, it passes
everywhere.

Bug: 1028283
Change-Id: I640f450e50df45c404b46c17c2d9cdce49313c60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4851233
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193796}
NOKEYCHECK=True
GitOrigin-RevId: 531ff1972a946a70d54102d3b759d63b5408f2f2