Consider nested snap areas on layout snaps
When snapping after a layout change, we snap to the
scroll-snap-align-specified position. This may cause unwanted jumps
in the case of large snap areas, e.g. when android shows browser
controls.
For large snap areas (areas larger than the snapport) we should,
instead, consider snapping to the closest covering position.
This is the behavior for snaps that occur at the end of a scroll.
This patch makes snaps after layout changes conform to this behavior.

Large snap areas may contain nested snap areas which may themselves be large snap areas that contain nested snap areas, so this patch
accounts for this by looking at all snap areas that overlap with the
snap area we are attempting to snap to.

A few notes about the test changes:

snap-area-overflow-boundary.html & changing-scroll-snap-align.html:
the first test cases now takes into account that the relayout
doesn't reset the scroll position.
A test case is also added for changing-scroll-snap-align.html verifying the behavior of small snap targets.

Bug: 1467300
Change-Id: I55d3a48838f701ca234cea2eb389daa70aadeacd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4842691
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193830}
NOKEYCHECK=True
GitOrigin-RevId: 46f6fa748b03540e0e8d7212964aefaac7cc3aec