[wallets] Rename MDoc* -> Wallet* for the WebIDL and the Mojo interfaces
As we explored, investigated and prototyped the API, we found that we
can deal with mdocs as a format, as opposed to a top-level API concept.

This CL has no external-facing consequence: the Web API is the same,
just the internal non-exposed dictionary names and Mojom classes.

Bug: 1416939
Change-Id: Iaca34c549cd0d7d51ad882eca98c98f8d65a3751
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4847891
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Rick Byers <rbyers@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Commit-Queue: Sam Goto <goto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1193853}
NOKEYCHECK=True
GitOrigin-RevId: 09fed5e42275f0470cb846a214c7b320c41fe21b