Renderer Process Crashes in Print Preview Popup Due to Null Return in GetFrame() During Execution.
If a web page has a button that triggers a print preview popup, and when the popup appears,
clicking the refresh button causes the Renderer process to crash.
If loading_for_print_ is true, GetFrame()->Client()->DispatchDidFinishLoadForPrinting() that after the execution is completed, there is a possibility of GetFrame() returning null.
So after this, it is necessary to determine whether it is empty.
If it is empty, false should be returned.

Bug: 1478487
Change-Id: Ieac77cba80f7c0115dfaea1c0dc3731a5b30dd2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4835376
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1194228}
NOKEYCHECK=True
GitOrigin-RevId: 568fa0dde09cf78bb079b21c9c0232ae0494c18d