Re-land: CHECK that style and layout are not left dirty during lifecycle
The first iteration of this patch, https://crrev.com/1192759, was
reverted due to scroll animations leaving style and layout dirty
(https://crbug.com/1478009). This turns out to be working as intended
so this patch moves the cleanliness check below
`RunScrollSnapshotClientSteps`.

Original change description below:
`LocalFrameView::UpdateLifecyclePhasesInternal` is responsible for
running the lifecycle and uses the following high-level pattern:
  while (true) {
    RunStyleAndLayoutLifecyclePhases();

    bool needs_to_repeat_lifecycle = RunStepA();
    if (needs_to_repeat_lifecycle) { continue; }

    bool needs_to_repeat_lifecycle = RunStepB();
    if (needs_to_repeat_lifecycle) { continue; }
    ...
  }
Each step expects clean style and layout, and if a step dirties style
or layout, it should re-start the lifecycle. This patch adds CHECKs for
many of the lifecycle steps which ensures style and layout are not left
dirty. These checks are added inside functions when possible, to
provide more helpful crash stacks.

Bug: 1479251, 1473823
Change-Id: I3679255751de6d22feb6bc3b14989a588352b06b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4850750
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1194398}
NOKEYCHECK=True
GitOrigin-RevId: 78dcd4cdcb498ed2074435d709f752f88af080a3