Reland "Allow anchor interaction tracking on localhost"
This is a reland of commit bf798ff87519d500f82b4238d4cb69dffaf67074

The original CL caused NavigationPredictorBrowserTest.PipelineHttp to
start failing, as the test uses a loopback address with http. The
reland updates the test to use another hostname which doesn't get
the special treatment that localhost does.

Furthermore, the original CL passed the CQ because it does not
reliably fail. The reland addresses this by waiting for the
navigation to finish instead of relying just on RunUntilIdle.

Original change's description:
> Allow anchor interaction tracking on localhost
>
> There's an explicit check that a document's URL is HTTPS before
> computing anchor element metrics. This notably excludes
> http://localhost. We now check for HTTP/HTTPS and potential
> trustworthiness of the HTTP case.
>
> Bug: None
> Change-Id: I6307808792d8ff60059f1de2c54287ad83e33d2e
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4846664
> Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
> Auto-Submit: Kevin McNee <mcnee@chromium.org>
> Reviewed-by: Simon Pelchat <spelchat@chromium.org>
> Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1193952}

Bug: None
Change-Id: Ia4f6bc4d918a222635463fcd30ad46ae69c992a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4852020
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Simon Pelchat <spelchat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1194902}
NOKEYCHECK=True
GitOrigin-RevId: 020bc90be051f8a85dbd5af7369083bd9a7e78b6