Move SubResourceAttributeName method from Element to HTMLFrameOwnerElement.
This moves Element::SubResourceAttributeName to HTMLFrameOwnerElement,
since its only caller has an HTMLFrameOwnerElement, which means the
other implementations are unused, and can be removed.  (Some of those
other implementations, such as on HTMLLinkElement, appear pretty buggy.)
This leaves a stub implementation on HTMLFrameOwnerElement with
overrides on HTMLEmbedElement and HTMLObjectElement.

(I ran into this while looking for the checks we do of the text/css MIME
type for https://crbug.com/1477672 .)

Change-Id: Ie55837ec87e7f794a5cea36e47e0dd8fef81ef73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4852855
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1194966}
NOKEYCHECK=True
GitOrigin-RevId: 841feab39883a6e8e29edff118cc4606d1c9b1d1