Use scrollend events to signal end-of-scroll
This -expected.txt file for snap-area-overflow-boundary.html was
recently removed. It is now flaking and it's probably because the
test's end-of-scroll signal is a timeout/counting frames.
Scrollend events are a more reliable signal.
As this is a WPT and not all browsers support scrollend, a fallback is
added for environments that don't support scrollend with an increased
frame count.

The TestExpectations entry will be removed after verifying that the
test is passing as expected.

Bug: 1480092,1480070,1480126,1400399
Change-Id: I24e50ec922ee5bf9007330c7926ff30bb234ef69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4852800
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1194988}
NOKEYCHECK=True
GitOrigin-RevId: 81843370d28cc29e4718af596b6934fa4db8c91d