Handle OOFs in spanners within NGOutOfFlowLayoutPart::ReplaceFragment
When we find an OOF in the CB path within
NGOutOfFlowLayoutPart::ReplaceFragment(), we end up looking for the
next viable fragmentation context root to search for its fragment(s).
However, if there is a spanner in the CB path (as a result of
CL:3343283), that OOF will propagate up to the next context since
spanners are not laid out in their immediate multicol parent.

Thus, we will need to continue up the tree in such cases in
NGOutOfFlowLayoutPart::ReplaceFragment() in order to properly find
the OOF ancestor in question.

Bug: 1479320
Change-Id: I88ffa3094f26c06cebc4bbe32a59e41ad77fe763
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4854084
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1195065}
NOKEYCHECK=True
GitOrigin-RevId: 663eaab09a86b923f5a73ebbb61da1358859210c