Disable side-effects when laying out grid item for space shortage.
When we lay out a grid item to figure out the block-size of its first
fragment, so that we can report space shortage for column balancing, be
sure to establish a scope for disabling side-effects, since we just drop
the resulting fragment on the floor after having checked its size.

I don't see how it would be possible to add an automated test for this
at the moment, because crbug.com/1481034 triggers a DCHECK failure
before we get to the accessibility code, which triggers the crash that
this bug is about.

Bug: 1480911
Change-Id: I4c3b0bac435b512394088afd616c05a56cdafc01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4853710
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1195080}
NOKEYCHECK=True
GitOrigin-RevId: 17c77cb612fd58489469fa1eb0fa49b443029ffa