[A11y] Don't allow creation of objects while tree is frozen
The line `CHECK(IsProcessingDeferredEvents())` in
AXObjectCacheImpl::CreateAndInit() is not intended to include tree
serialization time, and object creation should not occur then.

Fix it so that IsProcessingDeferredEvents() is set to false when it's
finished (right before the tree is frozen for serialization).

Bug: none
Change-Id: Ic058339864f269411aba66909bf76bc2bf6e1ac6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4851413
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1195139}
NOKEYCHECK=True
GitOrigin-RevId: 5d9a0faededa2d306ffdb4df35d2d5dea58897c3