blink: refactor Element::StyleAttributeChanged
This moves out the function for determining if the modification
is enabled into a standalone function. I originally did this as
I was hitting a crash in a test because the ContentSecurityPolicy
was null. This restructuring orders based on most likely things to
happen. There should be no functional change here.

Change-Id: If3f5d7da0b50ed5388a71ad6e9ea57d8308b7bf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4857359
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1195525}
NOKEYCHECK=True
GitOrigin-RevId: d1c9278a4a591320f137b07532d2548655c31b9c