Fix SW static routing API running status condition.
In the past implementation, we used the cached value for the running
status when it is asked inside the ServiceWorkerSubresourceLoader.
It is a bit stale value and the caller may get too old value to use.
This CL makes the value updated on ServiceWorkerVersion's running status
change to avoid such mismatch as much as possible.  For that purpose,
when the ServiceWorker static routing rules have rules with
running status condition, a mojo IPC receiver is created and passed
to the controller ServiceWorker connector to make the connector to be
the receiver of the running status.

We also fixed the test.  The test expects the ServiceWorker has been
stopped.  However, after `ReloadBlockUntilNavigationsComplete` is
called, the ServiceWorker awakes, and we need to explicitly stop it
to make the test work.

Bug: 1371756
Change-Id: I9ed5a6530c4fe33789fe794019e63d25c600d4bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4853777
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1195735}
NOKEYCHECK=True
GitOrigin-RevId: 882ed29c084f46b0a7f243921f534068d31376a0