Mark logical inset properties on boxes as layout dependent
After crrev.com/c/4583533, when getComputedStyle() gets an inset
property value from an out-of-flow box, it needs to check the box's
layout result. For sanity, it added a CHECK to assert that the box's
layout is already done.

This caused a crash because logical inset properties aren't currently
makred as layout-dependent, so we are not updating layout when getting
their values. This patch fixes it by marking them layout-dependent,
and implementing all the IsLayoutDependent() overrides in the same way
as the physical properties.

Fixed: 1478578
Change-Id: Iaa6e3380b692925e50801a237c28857c339a2e2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4857624
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Auto-Submit: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1195780}
NOKEYCHECK=True
GitOrigin-RevId: 7bb0b6845cb832a2f3070e0cc0d61368e2554969