Make MediaStreamTrack Stats API synchronous (WebIDL).
In accordance with
https://github.com/w3c/mediacapture-extensions/pull/106, we should
synchronously return the frame counters to JavaScript rather than
returning a promise that is resolved with the frame counters.

This CL updates WebIDL, blink implementation and WPTs accordingly.

The run-to-completion is still not implemented (caching if the getter
is called multiple times in the same task execution cycle), but that
is easier to add in a follow-up CL.

Bug: chromium:1472978
Change-Id: Ieaf8bc902864ddbc12fbb3b00225836e1750f30a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4858239
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Tony Herre <toprice@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1195916}
NOKEYCHECK=True
GitOrigin-RevId: 7f5fb08d7067adb5e7d88104913612c13a846732