Add {{}} syntax for NodeParts
This CL enables this syntax:

  <span parseparts>
    <div {{}} id=nodepart_here></div>
  </span>

The above snippet will now delete the {{}} and connect a NodePart to
the `<div>`.

This is implemented by adding a bit `needs_node_part` to `HTMLToken`
and `AtomicHTMLToken`. The tokenizer is updated to look for `{{}}`
while in the two states where this could be encountered while
tokenizing attribute names, and the `needs_node_part` bit is set.
This same bit is plumbed through `HTMLConstructionSiteTask`. If found
in `HTMLConstructionSite::ExecuteTask()`, `ConstructNodePart` adds a
`NodePart`.

Bug: 1453291
Change-Id: Ib7b2f34b39fdafd19816cbb4cdd93c292cd5c6fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4848471
Reviewed-by: David Baron <dbaron@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1196015}
NOKEYCHECK=True
GitOrigin-RevId: f34a60db7df130ff6aa8c5dc1ff11434ac2f84af