Pre-Paint: Enter childful but fragment item-less culled inline.
We missed it completely. The situation here was an AREA element with an
OOF child, so that the AREA element is a culled / special inline not
generating any fragment items, and the OOF child similarly doesn't
create any fragment items, and we'd miss the subtree completely.

Bug: 1480474
Change-Id: Ief10c1e0c206ed56c36e20a53ba9c4de63d7cd21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4859287
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1196259}
NOKEYCHECK=True
GitOrigin-RevId: 2d791391a69b5f96191c52dbe98e02042a71d1d2