Mark checking ComputeContentSize optimization expensive
This `DCHECK` is quite expensive, not intended to be enabled
on user machines, while still useful when making a big changes
to the inline algorithms such as adding a new property.

This patch marks it as `EXPENSIVE_DCHECKS_ARE_ON` so that it's
not enabled even when `DCHECK` is enabled.

Bug: 1362791
Change-Id: I00cccaa9859329b84e7ca955a6ea3e30904973de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4863020
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1196500}
NOKEYCHECK=True
GitOrigin-RevId: b7d0c0fca5f8a2f96262886f02a7d41a3a812a2c