Do not use the public method for looking up the MainThreadIsolate
Use the internal method for code inside blink. The plan is for
the public API to be renamed so it can only be used for testing. So
we need to convert the internal users to use the internal API and
then we will restrict that APIs access in the future.

Bug: 263412
Change-Id: Ib20ce43fb8d562dbc91003b167970bd7deadf4b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4863121
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1196554}
NOKEYCHECK=True
GitOrigin-RevId: b71a976a8feb0731cff8313e9d251b3212e3bdd2