Update and move old CSS2.1 counters scope tests to wpt
Test changes are due to the:
https://drafts.csswg.org/css-lists/#inheriting-counters
which says that we inherit the counter from sibling, only
if element counters does not already contain a counter
with the same name. Also, previous tests were
taking the value of the counter from the wrong source (counter with
same name, not same name + same creator).

Bug: 990657
Change-Id: I27390b03b32fad73b7c01ec6a521d173fa12be40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4839123
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@google.com>
Cr-Commit-Position: refs/heads/main@{#1196674}
NOKEYCHECK=True
GitOrigin-RevId: 2e71534de83f7f02995637eccb8beb00f6915dca