Disable overflow-menu-keyboard-navigation.html on Win
The test was enabled in a previous CL (crrev.com/c/4854729) that fixes
multiple test falkiness on media controls overflow menu. But it turns
out that this particular test was disabled due to a different issue and
it remains flaky. So this CL disables the test again on Win.

Bug: 1480468
Change-Id: I4b8707ceda181864cd3460220248775120fb227d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865096
Auto-Submit: Muyao Xu <muyaoxu@google.com>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: Takumi Fujimoto <takumif@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1196829}
NOKEYCHECK=True
GitOrigin-RevId: 526a8708fdf68eda5fa43b74e9f0817217966864