Reorder logic at end of MovePastBreakpoint().
One intentional change: Don't unconditionally move past a child fragment
that is taller than space left. The assumption was that if it breaks
inside, it doesn't overflow the fragmentainer, so we could just go ahead
and move past it. Breaking before instead if it doesn't fit should now
be safe, with crrev.com/c/4831191 in. This change will pave the way for
removing NGLayoutResult::kOutOfFragmentainerSpace from the column layout
algorithm, so that it's possible to break before a nested multicol
fragment that's too tall, even if it has break tokens inside (for the
inner fragmentation context). Making a column row in a nested multicol
container taller than what we have room for may be necessary when
there's monolithic content in there, if we want to get rid of
kOutOfFragmentainerSpace.

Change-Id: I704ab874b1bd6e95037a1f0cc5302945ccddeeaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4831296
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1196993}
NOKEYCHECK=True
GitOrigin-RevId: e8e3cf1e7188934a3c0c0f48745ff680bf4878a9