Put kOutOfFragmentainerSpace in the corner.
All the problems caused by this status made me run OutOfPatience.
Remove most of the kOutOfFragmentainerSpace usage.

Don't abort multicol container layout with kOutOfFragmentainerSpace.
This has caused trouble for all the layout algorithms over time, which
the algorithms have had to handle. This code can now be removed. Instead
of aborting when the column layout algorithm encounters content
overflowing a column (due to constrained space in an outer fragmentation
context), stretch the column to encompass overflow, and relayout the
row, even if that causes the outer fragmentainer to be overflowed.

Then, we can decide afterwards whether and where to insert a regular
break, just like we already do for all the other layout algorithms. If
the row has container separation, we can break before it. If not, some
algorithm further up will figure out where to insert an optimal break.

Split MovePastBreakpoint() in two, so that the column layout algorithm
can use it without providing an NGLayoutInputNode.

Ideally, kOutOfFragmentainerSpace should be removed completely, but we
still need it for BR clear=all. We should consider fixing that, too.

Although this fixes a crash in NGGridLayoutAlgorithm (by removing the
kOutOfFragmentainerSpace handling code there, which was probably correct
anyway), the underlying issue seems to be crbug.com/1477564 .

Bug: 1477322
Change-Id: I113964abd2528ec9445a53fdd4f324b1c74d5e8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4831811
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197002}
NOKEYCHECK=True
GitOrigin-RevId: ea07d9c099493e9117aa0656d364e331adf25f34