Don't skip spanners if a multicol
In CL:4854084, we updated NGOutOfFlowLayoutPart::ReplaceFragment such
that is an OOF is found, we continue up the tree if a spanner is found
in the CB chain since spanners are not laid out in the nearest multicol
ancestor. However, if the spanner is a multicol itself, it shouldn't
be skipped.

Bug: 1482514
Change-Id: If5a5f804696e15d6d22a09117dea703fd471da7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4867258
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1197144}
NOKEYCHECK=True
GitOrigin-RevId: 0cc291c47d68c06f5ce757be7af74367ae280fbe